feat: add type in entities
