+++ src/entities/index.ts
@@ -1,0 +1,4 @@
+export * from './comments/type'
+export * from './posts/type'
+export * from './users/type'
+export * from './tags/type'
