+++ src/entities/tags/type.ts
@@ -1,0 +1,7 @@
+export interface Tag {
+    slug: string
+    name: string
+    url: string
+  }
+  
+export type Tags = Array<Tag>
